modularize code and remove most input shape constraints
